✨ feat: 반응형 fetch 개선
