+++ package.json
@@ -12,7 +12,8 @@
     "axios": "^1.7.2",
     "next": "14.2.3",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "react-responsive": "^10.0.0"
   },
   "devDependencies": {
     "typescript": "^5",
